Make flow pages repeatable
