+++ src/components/FlowPage.css
@@ -1,0 +1,7 @@
+.Input {
+  text-align: left;
+  width: 100%;
+  padding: 2em 1.125em 1em 1.125em;
+  border: 2px solid #d9d9d9;
+  border-radius: 0.5em;
+}
