+++ src/components/Warning.css
@@ -1,11 +1,3 @@
-.Input {
-  text-align: left;
-  width: 100%;
-  padding: 2em 1.125em 1em 1.125em;
-  border: 2px solid #d9d9d9;
-  border-radius: 0.5em;
-}
-
 .Warning {
   display: flex;
   align-items: center;
